Split north into lib and bin
